requirement.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
deeputils==3.1.6
Deprecated==1.2.7
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
imageio==2.8.0
itypes==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.2
olefile==0.46
opencv-python==4.2.0.32
Pillow==7.0.0
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.23.0
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
wrapt==1.12.1
